requirements.txt
opencv-python
scikit-image
matplotlib
numpy
torchvision
albumentations
segmentation-models-pytorch